Move all docs to modules
