+++ crypto/module.md
@@ -1,1 +1,10 @@
 # Module crypto
+
+The crypto module contains classes and utilities that make it easier to work with Digital Signature Algorithms in a way
+that will be ready for enterprise use.
+
+# Package web5.sdk.crypto
+
+This package contains the `KeyManager` interface and implementations, which are used to manage keys for signing and
+encryption.
+
